salaa: Change default status bar height
AOSP has changed how the status bar height is determined
in commit [1]. Since then status_bar_height_portrait is
ignored in favor of getting the height dynamically based
on display cutouts. Devices are supposed to keep the
default value and let system handle the display cutout.
However, we do not have a display cutout but instead
change the status bar height to match our heavily rounded
corners, which was not taken into consideration in the
commit. Hence change the default status bar height
directly so that the status bar height matches the
rounded corners.

[1]: LineageOS/android_frameworks_base@16496cb

Change-Id: Iea7e7412f79bc8df8ab45b659c09cabb608c2a42